+++ src/estd/port/arch/atmelsam.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// TODO: Root around https://developer.arm.com/documentation/dui0774/g/chr1383660321827 as well
+
+#ifdef __SAMD21G18A__
+#define ESTD_MCU_ATMEL_SAM
+#define ESTD_MCU_ATMEL_SAMD
+#endif