+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# v0.3.1 - DDMMMYY
+# v0.3.1 - 27MAY22
 
 ## Added Features